gae_to_gcs_services/requirements.txt
Flask==1.0.2
google-cloud-bigquery==1.9.0
google-cloud-storage==1.14.0
